requirements.txt
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.46
